+++ src/lib/containers/index.tsx
@@ -1,6 +1,7 @@
 import Artist from "./artist"
 import Gene from "./gene"
 import Home from "./home"
+import Inbox from "./inbox"
 import MyAccount from "./my_account"
 import WorksForYou from "./works_for_you"
 
@@ -10,4 +11,5 @@ export default {
   Gene,
   WorksForYou,
   MyAccount,
+  Inbox,
 }